Rename atc.js to challenge.js
In accordance with GA naming standards.
